ref(backend): Merge potato sent message
